Morph the activity feed instead of rebuilding it every 2s
The Activity tab re-polls every 2s with hx-swap="innerHTML", which destroys
and recreates all 100 rows on every poll even when nothing changed — the
flicker, and why text selection is lost mid-read.

One attribute. The infrastructure was already here and already used: idiomorph
is loaded in base.html, hx-ext="morph" is on <body>, and #task-list has
swapped this way since #11. The feed simply never got the same treatment.

Proved by node IDENTITY, which is the only thing that distinguishes a morph
from a replace — both emit identical HTML. The test stashes a reference to a
live row, waits a poll cycle, and asserts the row in the DOM is still the SAME
element. Mutation-checked: reverting to innerHTML fails it.

My first attempt at that test stamped a data-* attribute and failed against a
WORKING morph, because idiomorph syncs attributes and strips one the incoming
markup lacks even when the node itself is preserved. Worth knowing before
writing another test of this shape.

The open-state tracker for the #84 disclosures is still required for exactly
that reason: the server renders <details> without `open`, so morph would close
an open one.

e2e fixtures move to test/e2e/conftest.py so the suite can hold more than one
file — test_modal_smoke.py was carrying them and is modal-specific.

#library-rows still uses plain innerHTML but isn't on a timer (load/refresh
events only), so it's far less visible; left alone deliberately.

Fixes #91
